requirements.txt
fastapi
uvicorn
pandas
neuralprophet
streamlit
psycopg2-binary